+++ lib/enums/e_visibility.dart
@@ -6,10 +6,9 @@
  * licenses restricting copying, distribution and decompilation.
  */
 
-export 'src/enum.dart';
-export 'src/extensions.dart';
-export 'src/ff.dart';
-export 'src/page.dart';
-export 'src/page_transition.dart';
-export 'src/pipe.dart';
-export 'src/timer.dart';
+
+enum EVisibility {
+  visible,
+  invisible,
+  gone
+}