+++ lib/src/page.dart
@@ -13,6 +13,7 @@ import 'dart:collection';
 
 import 'package:ffm/src/enum.dart';
 import 'package:ffm/src/ff.dart';
+import 'package:ffm/src/page_transition.dart';
 import 'package:flutter/material.dart';
 
 import 'extensions.dart';