+++ lib/src/page_transition.dart
@@ -6,9 +6,13 @@
  * licenses restricting copying, distribution and decompilation.
  */
 
+library ff.page_transition;
+
 import 'package:flutter/material.dart';
 
 import 'enum.dart';
+
+part 'page_transition_holder.dart';
 
 class FPageTransition<T> extends PageRouteBuilder<T> {
   /// Child for your next page
@@ -21,7 +25,8 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
   final Widget? childCurrent;
 
   /// Transition types
-  ///  fade,rightToLeft,leftToRight, upToDown,downToUp,scale,rotate,size,rightToLeftWithFade,leftToRightWithFade
+  /// - fade, rightToLeft, leftToRight, upToDown, downToUp
+  /// - scale, rotate, size, rightToLeftWithFade, leftToRightWithFade
   final FPageTransitionType type;
 
   /// Curves for transitions
@@ -50,6 +55,8 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
   // ignore: public_member_api_docs
   final bool isIos;
 
+  late FPageTransitionHolder holder;
+
   /// Page transition constructor. We can pass the next page as a child,
   FPageTransition({
     Key? key,
@@ -67,20 +74,18 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
     this.isIos = false,
     this.matchingBuilder = const CupertinoPageTransitionsBuilder(),
     RouteSettings? settings,
-  })  : assert(inheritTheme ? ctx != null : true,
-  "'ctx' cannot be null when 'inheritTheme' is true, set ctx: context"),
+  })  : assert(inheritTheme ? ctx != null : true, "'ctx' cannot be null when 'inheritTheme' is true, set ctx: context"),
         super(
-        pageBuilder: (BuildContext context, Animation<double> animation,
-            Animation<double> secondaryAnimation) {
-          return inheritTheme
-              ? InheritedTheme.captureAll(ctx!, child)
-              : child;
-        },
-        settings: settings,
-        maintainState: true,
-        opaque: isOpaque,
-        fullscreenDialog: isFullscreenDialog,
-      );
+          pageBuilder: (BuildContext context, Animation<double> animation, Animation<double> secondaryAnimation) {
+            return inheritTheme ? InheritedTheme.captureAll(ctx!, child) : child;
+          },
+          settings: settings,
+          maintainState: true,
+          opaque: isOpaque,
+          fullscreenDialog: isFullscreenDialog,
+        ) {
+    holder = FPageTransitionHolder._(pageTransition: () => this);
+  }
 
   @override
   Duration get transitionDuration => duration;
@@ -90,41 +95,57 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
   Duration get reverseTransitionDuration => reverseDuration;
 
   @override
-  Widget buildTransitions(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child) {
+  Widget buildTransitions(
+      BuildContext context, Animation<double> animation, Animation<double> secondaryAnimation, Widget child) {
+    if (!holder.animationEnabled) {
+      return child;
+    }
+
+    final animate = holder
+        ._buildAnimation(
+          context: () => context,
+          animation: () => animation,
+          secondaryAnimation: () => secondaryAnimation,
+          child: () => child,
+        )
+        .exec;
+
     switch (type) {
       case FPageTransitionType.theme:
-        return Theme.of(context).pageTransitionsTheme.buildTransitions(
-            this, context, animation, secondaryAnimation, child);
+        return Theme.of(context)
+            .pageTransitionsTheme
+            .buildTransitions(this, context, animation, secondaryAnimation, child);
 
       case FPageTransitionType.fade:
-        if (isIos) {
-          var fade = FadeTransition(opacity: animation, child: child);
-          return matchingBuilder.buildTransitions(
-              this, context, animation, secondaryAnimation, fade);
-        }
-        return FadeTransition(opacity: animation, child: child);
-        // ignore: dead_code
-        break;
-
-    /// PageTransitionType.rightToLeft which is the give us right to left transition
-      case FPageTransitionType.rightToLeft:
-        var slideTransition = SlideTransition(
-          position: Tween<Offset>(
-            begin: const Offset(1, 0),
-            end: Offset.zero,
-          ).animate(animation),
-          child: child,
+        return animate(
+          animation: FadeTransition(opacity: animation, child: child),
         );
-        if (isIos) {
-          return matchingBuilder.buildTransitions(
-              this, context, animation, secondaryAnimation, child);
-        }
-        return slideTransition;
-        // ignore: dead_code
-        break;
 
-    /// PageTransitionType.leftToRight which is the give us left to right transition
+      /// PageTransitionType.rightToLeft which is the give us right to left transition
+      case FPageTransitionType.rightToLeft:
+        // var slideTransition = SlideTransition(
+        //   position: Tween<Offset>(
+        //     begin: const Offset(1, 0),
+        //     end: Offset.zero,
+        //   ).animate(animation),
+        //   child: child,
+        // );
+
+        return animate(
+          animation: SlideTransition(
+            position: Tween<Offset>(
+              begin: const Offset(1, 0),
+              end: Offset.zero,
+            ).animate(animation),
+            child: child,
+          ),
+        );
+
+      // return !isIos
+      //     ? slideTransition
+      //     : matchingBuilder.buildTransitions(this, context, animation, secondaryAnimation, child);
+
+      /// PageTransitionType.leftToRight which is the give us left to right transition
       case FPageTransitionType.leftToRight:
         return SlideTransition(
           position: Tween<Offset>(
@@ -136,7 +157,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
         // ignore: dead_code
         break;
 
-    /// PageTransitionType.topToBottom which is the give us up to down transition
+      /// PageTransitionType.topToBottom which is the give us up to down transition
       case FPageTransitionType.topToBottom:
         if (isIos) {
           var topBottom = SlideTransition(
@@ -146,8 +167,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
             ).animate(animation),
             child: child,
           );
-          return matchingBuilder.buildTransitions(
-              this, context, animation, secondaryAnimation, topBottom);
+          return matchingBuilder.buildTransitions(this, context, animation, secondaryAnimation, topBottom);
         }
         return SlideTransition(
           position: Tween<Offset>(
@@ -159,7 +179,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
         // ignore: dead_code
         break;
 
-    /// PageTransitionType.downToUp which is the give us down to up transition
+      /// PageTransitionType.downToUp which is the give us down to up transition
       case FPageTransitionType.bottomToTop:
         return SlideTransition(
           position: Tween<Offset>(
@@ -171,7 +191,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
         // ignore: dead_code
         break;
 
-    /// PageTransitionType.scale which is the scale functionality for transition you can also use curve for this transition
+      /// PageTransitionType.scale which is the scale functionality for transition you can also use curve for this transition
       case FPageTransitionType.scale:
         assert(alignment != null, """
                 When using type "scale" you need argument: 'alignment'
@@ -182,8 +202,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
             scale: animation,
             child: child,
           );
-          return matchingBuilder.buildTransitions(
-              this, context, animation, secondaryAnimation, scale);
+          return matchingBuilder.buildTransitions(this, context, animation, secondaryAnimation, scale);
         }
         return ScaleTransition(
           alignment: alignment!,
@@ -200,7 +219,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
         // ignore: dead_code
         break;
 
-    /// PageTransitionType.rotate which is the rotate functionality for transition you can also use alignment for this transition
+      /// PageTransitionType.rotate which is the rotate functionality for transition you can also use alignment for this transition
       case FPageTransitionType.rotate:
         assert(alignment != null, """
                 When using type "RotationTransition" you need argument: 'alignment'
@@ -220,7 +239,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
         // ignore: dead_code
         break;
 
-    /// PageTransitionType.size which is the rotate functionality for transition you can also use curve for this transition
+      /// PageTransitionType.size which is the rotate functionality for transition you can also use curve for this transition
       case FPageTransitionType.size:
         assert(alignment != null, """
                 When using type "size" you need argument: 'alignment'
@@ -238,7 +257,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
         // ignore: dead_code
         break;
 
-    /// PageTransitionType.rightToLeftWithFade which is the fade functionality from right o left
+      /// PageTransitionType.rightToLeftWithFade which is the fade functionality from right o left
       case FPageTransitionType.rightToLeftWithFade:
         return SlideTransition(
           position: Tween<Offset>(
@@ -259,7 +278,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
         // ignore: dead_code
         break;
 
-    /// PageTransitionType.leftToRightWithFade which is the fade functionality from left o right with curve
+      /// PageTransitionType.leftToRightWithFade which is the fade functionality from left o right with curve
       case FPageTransitionType.leftToRightWithFade:
         return SlideTransition(
           position: Tween<Offset>(
@@ -545,7 +564,7 @@ class FPageTransition<T> extends PageRouteBuilder<T> {
         // ignore: dead_code
         break;
 
-    /// FadeTransitions which is the fade transition
+      /// FadeTransitions which is the fade transition
       default:
         return FadeTransition(opacity: animation, child: child);
     }