+++ lib/src/pipe.dart
@@ -10,8 +10,7 @@ library ff.pipe;
 
 import 'dart:async';
 
-import 'package:ffm/src/ff.dart';
-import 'package:ffm/src/page.dart';
+import 'package:ffm/ffm.dart';
 import 'package:flutter/material.dart';
 import 'package:rxdart/subjects.dart';
 
@@ -43,7 +42,7 @@ class FPipe<T> {
     return _errPipe!.value;
   }
 
-  final _subscriptionListeners = <void Function(T val)>[];
+  final _subscriptionListeners = <String, void Function(T val)>{};
   var subscriptionSkippedCount = 0;
   var _disposed = false;
 
@@ -132,25 +131,18 @@ class FPipe<T> {
     }
 
     if (value != val) {
-      if (textEditingController == null) {
-        update(val);
-      } else if (ff.func.isTypeOf<T, String>()) {
-        var text = value as String;
-        textEditingController!
-          ..text = text
-          ..selection = TextSelection.collapsed(offset: text.length);
-
-        update(val);
-      }
+      update(val);
     }
   }
 
-  void subscribe({required void Function(T val) listener, int skippedCount = 0}) {
+  void subscribe({required void Function(T val) listener, int skippedCount = 0, FDisposer? disposer}) {
     if (_disposed) {
       return;
     }
 
-    _subscriptionListeners.add(listener);
+    final id = _generateId();
+    _subscriptionListeners[id] = listener;
+    disposer?.register(() => _subscriptionListeners.remove(id));
     subscriptionSkippedCount = skippedCount < 0 ? 0 : skippedCount;
     _eventSubscription ??= _pipe.listen(_subscriptionEvent);
   }
@@ -168,7 +160,7 @@ class FPipe<T> {
     if (subscriptionSkippedCount > 0) {
       subscriptionSkippedCount--;
     } else {
-      for (var listener in _subscriptionListeners) {
+      for (var listener in _subscriptionListeners.values) {
         listener(value);
       }
     }
@@ -191,10 +183,12 @@ class FPipe<T> {
       return;
     }
 
-    for (var listener in _subscriptionListeners) {
+    for (var listener in _subscriptionListeners.values) {
       listener(value);
     }
   }
+
+  String _generateId() => '${ff.func.randomChar(5)}${DateTime.now().millisecondsSinceEpoch}';
 
   @protected
   void dispose() {
@@ -204,6 +198,7 @@ class FPipe<T> {
       textEditingController?.dispose();
       _pipe.close();
       _errPipe?.close();
+      _subscriptionListeners.clear();
     }
   }
 }